separate main App and demo route with react-router
