+++ client/index.jsx
@@ -1,5 +1,12 @@
 import React from 'react';
 import { render } from 'react-dom';
+import { HashRouter as Router } from 'react-router-dom';
+
 import App from './components/app.jsx';
 
-render(<App />, document.getElementById('root'));
+render(
+  <Router>
+    <App />
+  </Router>,
+  document.getElementById('root')
+);
