+++ package.json
@@ -21,7 +21,8 @@
     "graphql-tools": "^6.2.4",
     "json-server": "^0.16.2",
     "react": "^16.14.0",
-    "react-dom": "^16.14.0"
+    "react-dom": "^16.14.0",
+    "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.3",
